sizing-automation/requirements.txt
pymupdf>=1.25.1
python-docx>=1.1.2
openpyxl>=3.1.5
extract-msg>=0.52.0
